requirements.txt
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==5.0.1
bidict==0.23.1
blinker==1.8.2
cachelib==0.13.0
cbor2==5.6.5
certifi==2025.7.14
cffi==1.17.1
click==8.1.7
cryptography==45.0.5
dnspython==2.6.1
ecdsa==0.19.1
eventlet==0.36.1
exceptiongroup==1.3.0
Flask==3.0.3
Flask-Session==0.8.0
Flask-SocketIO==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ic-py==1.0.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
leb128==1.0.8
MarkupSafe==2.1.5
mnemonic==0.20
msgspec==0.18.6
packaging==24.1
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
redis==6.2.0
simple-websocket==1.0.0
six==1.17.0
sniffio==1.3.1
tomli==2.2.1
typing_extensions==4.14.0
waiter==1.5
Werkzeug==3.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
